Abstract method for parse caching and rename file
